Add block-level rendering capability
Allows for specific fill (color, image, or custom function) for
any individual block (specified in 'mapLocs', values of which
can be array for a color, object for a p5.Image or a function
that should return one of the prior two fill types).

With the aforementioned functionality, add rendering of the player's
avatar at block (0,0) - hardcoded for now - and a selection border
box around the currently-selected block.