Fix a race condition upon avatar creation
The client could have made a GET for the avatar contents
before the next engine tick, resulting in a null return. This
serializes those operation by returning to the synchronous
call model that was used prior to introducing engine.js

Also fixes the same race in world creation (didn't cause issue,
as they are handled differently at creation in http.js).

Still lots more to improve in these paths, but this is a critical
fix.